add error when user is already confirmed

+++ api/api/facades/user.rb
@@ -33,6 +33,7 @@ class UserFacade
     hash = Encryption.hmac(Duse.config.secret_key, token)
     token = Duse::Models::ConfirmationToken.find_by_token_hash(hash)
     fail Duse::NotFound if token.nil?
+    fail Duse::AlreadyConfirmed if token.user.confirmed?
     token.user.confirm!
     token.destroy
   end
@@ -40,6 +41,7 @@ class UserFacade
   def resend_confirmation!(email)
     user = Duse::Models::User.find_by_email email
     fail Duse::NotFound if user.nil?
+    fail Duse::AlreadyConfirmed if user.confirmed?
     Duse::Models::ConfirmationToken.delete_all(user: user)
     ConfirmationEmail.new(user).send
   end
